+++ packages/orq-rc/src/orq_ai_sdk/models/filelistop.py
@@ -74,7 +74,7 @@ class FileListData(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-08-27T12:47:54.349Z")
+    created: Optional[datetime] = parse_datetime("2025-08-28T06:20:02.598Z")
     r"""The date and time the resource was created"""