+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -86,7 +86,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K3NS48MXAG6JE5M1256E0NPC"
+        "01K3QNASDYM9M8A9DZZRBBHYRZ"
     )
     r"""The id of the resource"""