+++ packages/orq-rc/RELEASES.md
@@ -2419,3 +2419,13 @@ Based on:
 - [python v3.12.0-rc.18] packages/orq-rc
 ### Releases
 - [PyPI v3.12.0-rc.18] https://pypi.org/project/orq-ai-sdk/3.12.0-rc.18 - packages/orq-rc
+
+## 2025-08-28 06:21:43
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.606.8 (2.687.13) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.12.0-rc.19] packages/orq-rc
+### Releases
+- [PyPI v3.12.0-rc.19] https://pypi.org/project/orq-ai-sdk/3.12.0-rc.19 - packages/orq-rc